REF: move egg metadata handling in separate private module.
