+++ okonomiyaki/repositories/enpkg.py
@@ -5,14 +5,16 @@ import zipfile
 
 import os.path as op
 
-from ..bundled.traitlets import HasTraits, Bool, Enum, Float, \
-    Instance, List, Long, Unicode
-from ..file_formats.egg import Dependency, _decode_none_values, \
-    _encode_none_values, egg_name, info_from_z, split_egg_name
+from ..bundled.traitlets import (
+    HasTraits, Bool, Enum, Float, Instance, List, Long, Unicode
+)
+from ..file_formats.egg import (
+    Dependency, egg_name, info_from_z, split_egg_name
+)
 from ..file_formats.setuptools_egg import parse_filename
 from ..utils import compute_md5
+from ..utils.traitlets import NoneOrUnicode
 
-_CAN_BE_NONE_KEYS = ["osdist", "platform", "python"]
 
 _AVAILABLE_DEFAULT = False
 _PRODUCT_DEFAULT = "commercial"
@@ -34,7 +36,7 @@ class EnpkgS3IndexEntry(HasTraits):
     egg_basename = Unicode()
     packages = List(Instance(Dependency))
     product = Enum(["pypi", "commercial", "free"], _PRODUCT_DEFAULT)
-    python = Unicode()
+    python = NoneOrUnicode()
     size = Long()
     type = Enum(["egg"], "egg")
     version = Unicode()
@@ -47,7 +49,6 @@ class EnpkgS3IndexEntry(HasTraits):
 
         Note: the passed in dictionary may be modified.
         """
-        data = _decode_none_values(data, _CAN_BE_NONE_KEYS)
         data["packages"] = [
             Dependency.from_spec_string(s) for s in data.get("packages", [])
         ]
@@ -125,7 +126,6 @@ class EnpkgS3IndexEntry(HasTraits):
             "type": self.type,
             "version": self.version,
         }
-        data = _encode_none_values(data, _CAN_BE_NONE_KEYS)
         return data
 
     def to_dict(self):
@@ -141,7 +141,6 @@ class EnpkgS3IndexEntry(HasTraits):
                 "size": self.size,
                 "type": self.type,
                 "version": self.version}
-        data = _encode_none_values(data, _CAN_BE_NONE_KEYS)
         return data
 
     def to_json(self):
